update file path to reflect project name
