+++ src/SftpFileSystemx64/FailedKeyDialog.cs
@@ -1,0 +1,44 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Drawing;
+using System.Text;
+using System.Windows.Forms;
+
+namespace SftpFileSystem
+{
+    public partial class FailedKeyDialog : Form
+    {
+        #region cTor
+
+        public FailedKeyDialog()
+        {
+            InitializeComponent();
+        }
+
+        #endregion
+
+        #region Events handler
+
+        private void button1_Click(object sender, EventArgs e)
+        {
+            DialogResult = DialogResult.OK;
+            Close();
+        }
+
+        private void button2_Click(object sender, EventArgs e)
+        {
+            DialogResult = DialogResult.Cancel;
+            Close();
+        }
+
+        private void button3_Click(object sender, EventArgs e)
+        {
+            DialogResult = DialogResult.Retry;
+            Close();
+        }
+
+        #endregion
+    }
+}
