+++ src/SftpFileSystemx64/KeyType.cs
@@ -1,0 +1,8 @@
+﻿namespace SftpFileSystem
+{
+    public enum KeyType
+    {
+        Putty,
+        Ssh
+    }
+}
