+++ src/SftpFileSystemx64/Properties/AssemblyInfo.cs
@@ -1,0 +1,9 @@
+﻿using System.Reflection;
+using System.Runtime.InteropServices;
+
+[assembly: AssemblyTitle("SftpFileSystem")]
+[assembly: AssemblyDescription("")]
+[assembly: AssemblyCulture("")]
+
+// The following GUID is for the ID of the typelib if this project is exposed to COM
+[assembly: Guid("6d67894f-87fd-40ad-bbce-0923fd38d0ff")]
