--- src/sftp-plugin/ConfigData.cs
@@ -1,15 +1,0 @@
-﻿namespace SftpFileSystem
-{
-    public class ConfigData
-    {
-        #region Properties
-
-        public string KeyFile { get; set; }
-
-        public bool UseKeyfile { get; set; }
-
-        public KeyType KeyType { get; set; }
-
-        #endregion
-    }
-}
