--- src/sftp-plugin/KeyType.cs
@@ -1,8 +1,0 @@
-﻿namespace SftpFileSystem
-{
-    public enum KeyType
-    {
-        Putty,
-        Ssh
-    }
-}
